+++ spec/rubocop/cop/performance/map_compact_spec.rb
@@ -46,6 +46,46 @@ RSpec.describe RuboCop::Cop::Performance::MapCompact, :config do
       RUBY
     end
 
+    it 'registers an offense when using `map.compact.first` with single-line method calls' do
+      expect_offense(<<~RUBY)
+        collection.map { |item| item.do_something }.compact.first
+                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `filter_map` instead.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        collection.filter_map { |item| item.do_something }.first
+      RUBY
+    end
+
+    it 'registers an offense when using `map.compact.first` with multi-line leading dot method calls' do
+      expect_offense(<<~RUBY)
+        collection
+          .map { |item| item.do_something }
+           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `filter_map` instead.
+          .compact
+          .first
+      RUBY
+
+      expect_correction(<<~RUBY)
+        collection
+          .filter_map { |item| item.do_something }
+          .first
+      RUBY
+    end
+
+    it 'registers an offense when using `map.compact.first` and there is a line break after `map.compact`' do
+      expect_offense(<<~RUBY)
+        collection.map { |item| item.do_something }.compact
+                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Use `filter_map` instead.
+          .first
+      RUBY
+
+      expect_correction(<<~RUBY)
+        collection.filter_map { |item| item.do_something }
+          .first
+      RUBY
+    end
+
     it 'does not register an offense when using `collection.map(&:do_something).compact!`' do
       expect_no_offenses(<<~RUBY)
         collection.map(&:do_something).compact!